# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Monster Realms Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/guest:
    post:
      operationId: guestLogin
      tags: [auth]
      summary: Create a guest account and get a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestLoginInput"
      responses:
        "200":
          description: Guest session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: registerPlayer
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username already taken
  /auth/login:
    post:
      operationId: loginPlayer
      tags: [auth]
      summary: Log in with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the authenticated player
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "401":
          description: Not authenticated

  # PLAYERS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GuestLoginInput:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 20
        avatarColor:
          type: string
      required:
        - username
    RegisterInput:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 20
        password:
          type: string
          minLength: 6
        email:
          type: ["string", "null"]
        avatarColor:
          type: string
      required:
        - username
        - password
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        player:
          $ref: "#/components/schemas/Player"
      required:
        - token
        - player

    # PLAYER
    Player:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatarColor:
          type: string
        isGuest:
          type: boolean
        explorerRank:
          type: string
        explorerXp:
          type: number
        explorerLevel:
          type: number
        tilesExplored:
          type: number
        energy:
          type: number
        maxEnergy:
          type: number
        coins:
          type: number
        monstersDiscovered:
          type: number
        monstersCaptured:
          type: number
        regionId:
          type: ["string", "null"]
        posX:
          type: number
        posY:
          type: number
        createdAt:
          type: string
      required:
        - id
        - username
        - avatarColor
        - isGuest
        - explorerRank
        - explorerXp
        - explorerLevel
        - tilesExplored
        - energy
        - maxEnergy
        - coins
        - monstersDiscovered
        - monstersCaptured
        - posX
        - posY
        - createdAt
...